GradioLearn/pyproject.toml
[project]
name = "gradiolearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.41.0",
    "numpy>=2.3.2",
    "openai>=1.99.3",
    "openai-agents>=0.2.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
